refactor: rename extend-res to extend-req
