+++ src/server/lib/extend-res.js
